test: add integration tests for resource market functionality
